Game Event Manager + Map Selection Menu
I used a placeholder image for the maps I know the highlighted colors are ugly but I need to find something that matches with the color palette. Also I made a custom game event manager for particle effects because I thought I needed a particle effect for when ur hovering over a button but its optional. I will probably do a demo on it or something down the line. ALSO I'll try and use the game event manager to load the scenes on click but I should probably just do a monobehaviour class. I just don't wanna reference a script in the scene.